requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
python-clamd
python-dotenv==1.0.0
